requirements.txt
# https://github.com/django/django
django>=2.1.2

# https://github.com/theskumar/python-dotenv
python-dotenv==0.7.1

# https://github.com/psycopg/psycopg2
psycopg2==2.7.4

# https://github.com/encode/django-rest-framework
djangorestframework==3.8.2

# https://github.com/jazzband/django-debug-toolbar
django-debug-toolbar==1.11

# https://github.com/carltongibson/django-filter
django-filter==2.0.0

# https://github.com/stefanfoulis/django-phonenumber-field
django-phonenumber-field==2.1.0
phonenumbers==8.10.

# https://github.com/alanjds/drf-nested-routers/
drf-nested-routers==0.90.0